Weather, first pass
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@3529 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ wilderness.c
@@ -211,7 +211,7 @@ int get_weather(int x, int y) {
   double trans_x;
   double trans_y;
   double result;
-  int    time_base;
+  double time_base;
   time_t now;
